+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - License information
 
+### Fixed
+- Pattern doesn't work
+
 ## [0.4.1]
 ### Added
 - `.editorconfig`