+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.5.0]
 ### Added
 - support for ellipse shape
 
@@ -161,7 +163,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Add Artboard parser
 
-[Unreleased]: https://github.com/L2jLiga/xd2svg/compare/v0.4.2...HEAD
+[Unreleased]: https://github.com/L2jLiga/xd2svg/compare/v0.5.0...HEAD
+[0.5.0]: https://github.com/L2jLiga/xd2svg/compare/v0.4.2...v0.5.0
 [0.4.2]: https://github.com/L2jLiga/xd2svg/compare/v0.4.1...v0.4.2
 [0.4.1]: https://github.com/L2jLiga/xd2svg/compare/v0.4.0...v0.4.1
 [0.4.0]: https://github.com/L2jLiga/xd2svg/compare/v0.3.1...v0.4.0